rename the ad binary to nad, GitHub #2337
The new name, nad, stands for Netatalk AppleDouble, emphasizing that it
is operating on Netatalk style AppleDouble metadata specifically,
and depends on Netatalk running on the same host that you run nad on.

This also reduces some ambiguity in the Netatalk codebase, since "ad"
is used internally to also refer to the AppleDouble module in libatalk,
for instance the "logtype_ad" logtype in the netatalk logger module.
