+++ bin/nad/ftw.c
