+++ bin/nad/nad_set.c
@@ -34,7 +34,7 @@
 #include <atalk/adouble.h>
 #include <atalk/cnid.h>
 
-#include "ad.h"
+#include "nad.h"
 
 #define ADv2_DIRNAME ".AppleDouble"
 
@@ -62,7 +62,7 @@ static char *new_attributes;
 static void usage_set(void)
 {
     printf(
-        "Usage: ad set [-t TYPE] [-c CREATOR] [-l label] [-f flags] [-a attributes] file|dir \n\n"
+        "Usage: nad set [-t TYPE] [-c CREATOR] [-l label] [-f flags] [-a attributes] file|dir \n\n"
         "     Color Label:\n"
         "       none | grey | green | violet | blue | yellow | red | orange\n\n"
         "     FinderFlags:\n"
