+++ doc/po4a.cfg
@@ -3,7 +3,6 @@
 
 [options] markdown --master-charset UTF-8
 
-[type: text] manpages/man1/ad.1.md $lang:translated/$lang/ad.1.md
 [type: text] manpages/man1/addump.1.md $lang:translated/$lang/addump.1.md
 [type: text] manpages/man1/aecho.1.md $lang:translated/$lang/aecho.1.md
 [type: text] manpages/man1/afp_lantest.1.md $lang:translated/$lang/afp_lantest.1.md
@@ -19,6 +18,7 @@
 [type: text] manpages/man1/fce_listen.1.md $lang:translated/$lang/fce_listen.1.md
 [type: text] manpages/man1/getzones.1.md $lang:translated/$lang/getzones.1.md
 [type: text] manpages/man1/macusers.1.md $lang:translated/$lang/macusers.1.md
+[type: text] manpages/man1/nad.1.md $lang:translated/$lang/nad.1.md
 [type: text] manpages/man1/nbplkup.1.md $lang:translated/$lang/nbplkup.1.md
 [type: text] manpages/man1/nbprgstr.1.md $lang:translated/$lang/nbprgstr.1.md
 [type: text] manpages/man1/pap.1.md $lang:translated/$lang/pap.1.md
